Revert "modifies SegUI WIP"
This reverts commit 00751dabbd18df6af6f5e22535d88470127e72b9.
